% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHGT.R
\name{minHGT}
\alias{minHGT}
\title{minHGT permutation test}
\usage{
minHGT(thisVector, numPerms = 1000)
}
\arguments{
\item{thisVector}{A vector of individual Bernoulli trials (https://en.wikipedia.org/wiki/Bernoulli_trial). Each value must be a 0 or 1.}
}
\description{
This function takes a list of indicator values (0s and 1s) ranked  by some orthogonal measure,
for example if you sorted people by height and then went down the row asking if each individual
was male (1) or female (0)
}
\details{
This permutation test determines if the successes in that ranked list are predisposed beyond
random chance (as in the case with height and sex, which are associated) to be skewed to the front of the list

More specifically, this test identifies the minimum hypergeometric test (minHGT) p-value for a given vector as described here:
https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-10-48
then performs a specified number of permutations of the list to determine how frequently an equal or lesser minHGT p-value can be obtained randomly.
the reported p-value from this script, is then simply just n/N, where n is the number of randomly obtained minHGT scores which were equal to or less
than the value obtained by the given list, then divided by the total number of permutations attempted.
}
\examples{
v1 <- round(runif(100,0,.7),0)
v2 <- round(runif(100,0.3,1),0)
testVector <- c(v2,v1)
testOut <- minHGT(testVector)
testOut$pval  # the probability of observing a given minimum mypergeometric value given random shuffling of the vector
testOut$crit  # the set size which yeilds the greatest minHGT for the given vector
}
